Linus fix

+++ Source/Network/BaseNetwork.cpp
@@ -156,13 +156,13 @@ void BaseNetwork::Update(float _dt)
 	}
 }
 
-DWORD BaseNetwork::GetMillisecondsTime()
+unsigned int BaseNetwork::GetMillisecondsTime()
 {
 #ifdef WIN32
-	return GetTickCount();
+	return (unsigned int)GetTickCount();
 #else
 	struct timeval tv;
 	if (gettimeofday(&tv, 0) != 0) return 0;
-	return (float)((tv.tv_sec * 1000ul) + (tv.tv_usec / 1000ul));
+	return (unsigned int)((tv.tv_sec * 1000ul) + (tv.tv_usec / 1000ul));
 #endif
 }
